[Chore] 글쓰기 프래그먼트 패키지 분리
