--- app/src/main/java/com/android/hara/presentation/write/fragment/ProsconsFragViewModel.kt
@@ -1,4 +1,0 @@
-package com.android.hara.presentation.write.fragment
-
-class ProsconsFragViewModel {
-}
