+++ app/src/main/java/com/android/hara/presentation/write/fragment/content/ContentFragViewModel.kt
@@ -1,11 +1,9 @@
-package com.android.hara.presentation.write.fragment
+package com.android.hara.presentation.write.fragment.content
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class TitleFragViewModel : ViewModel() {
+class ContentFragViewModel : ViewModel() {
     private val _answer = MutableLiveData<String>()
-
-    // 양방향데이터바인딩은 LiveData 사용 불가
     val answer get() = _answer
 }
