+++ app/src/main/java/com/android/hara/presentation/write/fragment/option/OptionFragViewModel.kt
@@ -1,0 +1,4 @@
+package com.android.hara.presentation.write.fragment.option
+
+class OptionFragViewModel {
+}
