+++ app/src/main/java/com/android/hara/presentation/write/fragment/option/WriteOptionFragment.kt
@@ -1,4 +1,4 @@
-package com.android.hara.presentation.write.fragment
+package com.android.hara.presentation.write.fragment.option
 
 import android.os.Bundle
 import android.view.View
@@ -6,12 +6,12 @@ import androidx.fragment.app.activityViewModels
 import androidx.navigation.NavController
 import androidx.navigation.Navigation
 import com.android.hara.R
-import com.android.hara.databinding.FragmentWriteProsconsBinding
+import com.android.hara.databinding.FragmentWriteOptionBinding
 import com.android.hara.presentation.base.BindingFragment
 import com.android.hara.presentation.write.WriteViewModel
 
-class WriteProsconsFragment :
-    BindingFragment<FragmentWriteProsconsBinding>(R.layout.fragment_write_proscons) {
+class WriteOptionFragment :
+    BindingFragment<FragmentWriteOptionBinding>(R.layout.fragment_write_option) {
     lateinit var navController: NavController
     private val writeViewModel: WriteViewModel by activityViewModels()
 
@@ -27,14 +27,14 @@ class WriteProsconsFragment :
     }
 
     private fun onClickNextBtn() {
-        binding.ibWriteProsconsNextButton.setOnClickListener {
-            navController.navigate(R.id.action_writeProsconsFragment_to_writeCategoryFragment)
+        binding.ibWriteOptionNextButton.setOnClickListener {
+            navController.navigate(R.id.action_writeOptionFragment_to_writeProsconsFragment)
             writeViewModel.addProgress()
         }
     }
 
     private fun onClickBackBtn() {
-        binding.ibWriteProsconsBackButton.setOnClickListener {
+        binding.ibWriteOptionBackButton.setOnClickListener {
             navController.navigateUp()
             writeViewModel.subProgress()
         }
