+++ app/src/main/java/com/android/hara/presentation/write/fragment/proscons/ProsconsFragViewModel.kt
@@ -1,0 +1,4 @@
+package com.android.hara.presentation.write.fragment.proscons
+
+class ProsconsFragViewModel {
+}
