+++ app/src/main/res/layout/fragment_write_title.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="vm"
-            type="com.android.hara.presentation.write.fragment.TitleFragViewModel" />
+            type="com.android.hara.presentation.write.fragment.title.TitleFragViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
